extract cors and test it
