+++ apps/meteor/.mocharc.js
@@ -36,7 +36,7 @@ module.exports = {
 		'tests/unit/lib/**/*.spec.ts',
 		'tests/unit/server/**/*.tests.ts',
 		'tests/unit/server/**/*.spec.ts',
-		'app/api/**/*.spec.ts',
+		'app/api/lib/**/*.spec.ts',
 		'app/file-upload/server/**/*.spec.ts',
 	],
 };
