+++ apps/meteor/jest.config.ts
@@ -38,6 +38,8 @@ export default {
 				'<rootDir>/ee/server/patches/**/*.spec.ts',
 				'<rootDir>/app/cloud/server/functions/supportedVersionsToken/**.spec.ts',
 				'<rootDir>/app/utils/lib/**.spec.ts',
+				'<rootDir>/app/api/server/**.spec.ts',
+				'<rootDir>/app/api/server/middlewares/**.spec.ts',
 			],
 		},
 	],
